feat(TCOMP-773): extend http api to handle generic use cases
refactor the HttpClientFactoryImpl
